power chart

+++ pkg/conversion.go
@@ -2,6 +2,6 @@ package pkg
 
 import "math"
 
-func DivisionPowerOfTen(num int64, power int) float64 {
-	return float64(num) / math.Pow(10, float64(power))
+func DivisionPowerOfTen(num float64, power int) float64 {
+	return num / math.Pow(10, float64(power))
 }
